+++ Meshkraft/build.gradle
@@ -56,15 +56,15 @@ afterEvaluate {
         publications {
             release(MavenPublication) {
                 from components.release
-                groupId = 'com.mertgolcu.meshkraft'
+                groupId = 'com.artlabs.meshkraft'
                 artifactId = 'meshkraft'
-                version = '1.2.0'
+                version = '1.2.1'
             }
             debug(MavenPublication) {
                 from components.debug
-                groupId = 'com.mertgolcu.meshkraft'
+                groupId = 'com.artlabs.meshkraft'
                 artifactId = 'meshkraft-debug'
-                version = '1.2.0'
+                version = '1.2.1'
             }
         }
         repositories {
