--- Meshkraft/src/main/java/com/mertgolcu/meshkraft/Constants.kt
@@ -1,17 +1,0 @@
-package com.mertgolcu.meshkraft
-
-/**
- * @author Mert Gölcü
- * @date 10.09.2021
- */
-
-
-const val VIEWER_URL = "https://arvr.google.com/scene-viewer/1.0"
-const val TITLE = "title"
-const val MODE = "mode"
-const val FILE = "file"
-const val FILE_TYPE = "glb"
-const val STANDARD_PACKAGE = "com.google.android.googlequicksearchbox"
-const val AR_ONLY_PACKAGE = "com.google.ar.core"
-const val TOKEN = "token"
-const val BASE_URL = "https://api.artlabs.ai/"
