--- Meshkraft/src/main/java/com/mertgolcu/meshkraft/MeshkraftBase.kt
@@ -1,15 +1,0 @@
-package com.mertgolcu.meshkraft
-
-/**
- * @author Mert Gölcü
- * @date 11.09.2021
- */
-class MeshkraftBase() {
-    val test = "test"
-
-    companion object {
-        fun getInstance(): Meshkraft {
-            return Meshkraft
-        }
-    }
-}
