+++ app/src/main/java/com/mertgolcu/testapp/MainActivity.kt
@@ -2,11 +2,11 @@ package com.mertgolcu.testapp
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.mertgolcu.meshkraft.Meshkraft
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-
     }
 }
