--- app/src/main/java/com/mertgolcu/testapp/MyApplication.kt
@@ -1,15 +1,0 @@
-package com.mertgolcu.testapp
-
-import android.app.Application
-
-/**
- * @author Mert Gölcü
- * @date 13.09.2021
- */
-//class MyApplication:Application() {
-//    override fun onCreate() {
-//        super.onCreate()
-//        Mashkraft.setApiKey("TOKEN_HERE")
-//    }
-//
-//}
