+++ build.gradle
@@ -2,7 +2,7 @@
 buildscript {
     repositories {
         google()
-        maven { url 'https://jitpack.io' }
+
     }
     dependencies {
         classpath "com.android.tools.build:gradle:7.0.2"
